+++ src/module/database/bangumi.py
@@ -13,34 +13,7 @@ class BangumiDatabase(DataConnector):
     def update_table(self):
         table_name = "bangumi"
         db_data = self.__data_to_db(BangumiData())
-        columns = ", ".join([f"{key} {self.__python_to_sqlite_type(value)}" for key, value in db_data.items()])
-        create_table_sql = f"CREATE TABLE IF NOT EXISTS {table_name} ({columns});"
-        self._cursor.execute(create_table_sql)
-        self._cursor.execute(f"PRAGMA table_info({table_name})")
-        existing_columns = {column_info[1]: column_info for column_info in self._cursor.fetchall()}
-        for key, value in db_data.items():
-            if key not in existing_columns:
-                add_column_sql = f"ALTER TABLE {table_name} ADD COLUMN {key} {self.__python_to_sqlite_type(value)} DEFAULT {value};"
-                self._cursor.execute(add_column_sql)
-        self._conn.commit()
-        logger.debug("Create / Update table bangumi.")
-
-    @staticmethod
-    def __python_to_sqlite_type(value) -> str:
-        if isinstance(value, int):
-            return "INTEGER NOT NULL"
-        elif isinstance(value, float):
-            return "REAL NOT NULL"
-        elif isinstance(value, str):
-            return "TEXT NOT NULL"
-        elif isinstance(value, bool):
-            return "INTEGER NOT NULL"
-        elif isinstance(value, list):
-            return "TEXT NOT NULL"
-        elif value is None:
-            return "TEXT"
-        else:
-            raise ValueError(f"Unsupported data type: {type(value)}")
+        self._update_table(table_name, db_data)
 
     @staticmethod
     def __data_to_db(data: BangumiData) -> dict:
